Add a note for date formatting.
LocalDate assumes a 2000 base for years, which means yy is parsed as 20yy. This assumption is probably incorrect in our use case given that the dates are implied to be birthdays and they are far in the future. I linked a stackoverflow questions which explains this and a possible solution.